Support deploying the site with a URL prefix
The original landscape assumed that the site would be deployed at the server
root, e.g. https://foo.bar.baz/

However, due to NREL messaging constraints and our lack of budget, we need to
deploy this on Github Pages. So the URL will be
https://foo.github.io/mobility_landscape

This is a problem because the router doesn't have the basename, so this
automatically redirects to https://foo.github.io/ and breaks all asset loading.

Fixes:
- make the basename configurable (in `src/store/configureStore.js`)
- configure the basename correctly for dev and prod (`webpack.config.dev.js` and `webpack.config.prod.js`)
- ensure that the dist files are packaged in `dist/mobility_net` so everything
  has a prefix (`webpack.config.prod.js` and `package.json`)
- change the `distServer` to serve files with a prefix so we can test out the
  final configuration before publishing

TODO: make the three locations:
- `package.json`
- `webpack.config.prod.js`
- `distServer.js`
read the prefix from an environment variable instead of hardcoding it